pyproject.toml
[project]
name = "agentic-dm"
version = "0.1.0"
description = "AI-powered D&D Dungeon Master Assistant"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.2.1",
    # LLM & Embeddings
    "openai>=2.6.1",
    "tiktoken>=0.8.0",
    # Vector Database
    "chromadb>=0.5.0",
    # Graph Database
    "neo4j>=6.0.2",
    # PDF Processing
    "pymupdf>=1.25.0",
    # NER & NLP
    "spacy>=3.8.0",
    "rapidfuzz>=3.0.0",
    "pyahocorasick>=2.0.0",
    # Agent Protocol
    "mcp>=1.19.0",
    # Utilities
    "pyyaml>=6.0.3",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "en-core-web-sm",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
ingest-pdf = "backend.scripts.ingest_pdf:main"
process-transcript = "backend.scripts.process_transcript:main"

[tool.setuptools.packages]
find = { where = ["."], include = ["backend*"] }

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
